refactor: reorganiza estrutura de pacotes do projeto
move a classe `Config` para o pacote `config` por ser uma configuração geral do jogo.
subdivide o pacote `model` em subpacotes mais específicos:
- `enums`: para enumeradores relacionados a entidades
- `interfaces`: para interfaces implementadas por entidades
- `entidades`: para classes de principais entidades do jogo
  - `inimigos`: para classes dos zumbis e outros inimigos
  - `plantas`: para classes de plantas que o jogador pode usar
